network-automation_py/device-check/requirements.txt
bcrypt==3.2.0
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
ciscoconfparse==1.5.19
colorama==0.4.3
cryptography==3.1
dataclasses==0.7
dnspython==2.0.0
future==0.18.2
idna==2.10
importlib-resources==3.0.0
Jinja2==2.11.2
junos-eznc==2.5.3
lxml==4.5.2
MarkupSafe==1.1.1
mypy-extensions==0.4.3
ncclient==0.6.9
netaddr==0.8.0
netmiko==3.3.0
ntc-templates==1.5.0
paramiko==2.7.2
passlib==1.7.2
pkg-resources==0.0.0
pycparser==2.20
pyeapi==0.8.3
PyNaCl==1.4.0
pyparsing==2.4.7
pyserial==3.4
PyYAML==5.3.1
requests==2.24.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
scp==0.13.2
six==1.15.0
tabulate==0.8.7
textfsm==1.1.0
transitions==0.8.2
typing-extensions==3.7.4.3
urllib3==1.25.10
yamlordereddictloader==0.4.0
zipp==3.1.0
